requirements.txt
django
django-multitenant
psycopg2
gunicorn
djangorestframework
djangorestframework-simplejwt